br/stream: move store related stuffs into store mod
